refactor: separate util functions
